Revamped Homepage
I have uploaded icons for each services. I have updated the details for the footer section across all pages. I have changed the Homepage header 1 to "Skin and Clear" and header 2 to "feel Confident, be confident". The headers are within place in desktop scale but goes out of view for mobile resolution. Utkarsh - Needs to update this.